tests: update test cases
